@@ -3,9 +3,10 @@
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "serve": "vue-cli-service serve",
-    "build": "vue-cli-service build",
-    "lint": "vue-cli-service lint",
+    "dev": "vite",
+    "build": "vite build",
+    "serve": "vite preview",
+    "lint": "eslint --ext .js,.vue --ignore-path .gitignore --fix src",
     "db:seed": "firebase firestore:delete --all-collections -f -r && node firestoreImport"
   },
   "dependencies": {
@@ -15,7 +16,7 @@
     "@hennge/vue3-pagination": "^1.0.17",
     "@vee-validate/i18n": "^4.7.4",
     "@vee-validate/rules": "^4.7.4",
-    "core-js": "^3.6.5",
+    "@vueuse/head": "^1.1.9",
     "dayjs": "^1.11.7",
     "firebase": "^8.10.1",
     "lodash": "^4.17.21",
@@ -27,20 +28,17 @@
     "vue-router": "^4.1.6"
   },
   "devDependencies": {
-    "@vue/cli-plugin-babel": "~4.5.9",
-    "@vue/cli-plugin-eslint": "~4.5.9",
-    "@vue/cli-service": "~4.5.9",
+    "@vitejs/plugin-vue": "^4.0.0",
     "@vue/compiler-sfc": "^3.0.0",
     "@vue/eslint-config-standard": "^5.1.2",
-    "babel-eslint": "^10.1.0",
-    "eslint": "^6.7.2",
+    "eslint": "^8.35.0",
     "eslint-plugin-import": "^2.20.2",
     "eslint-plugin-node": "^11.1.0",
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-standard": "^4.0.0",
-    "eslint-plugin-vue": "^7.0.0",
+    "eslint-plugin-vue": "^9.9.0",
     "firestore-export-import": "^1.3.5",
-    "vue-cli-plugin-webpack-bundle-analyzer": "~4.0.0"
+    "vite": "^4.1.4"
   },
   "browserslist": [
     "> 1%",
