@@ -1,16 +1,13 @@
 module.exports = {
   root: true,
   env: {
-    node: true
+    es2021: true
   },
   extends: [
     'plugin:vue/vue3-essential',
     '@vue/standard'
   ],
-  parserOptions: {
-    parser: 'babel-eslint'
-  },
   rules: {
-    'no-unused-vars': process.env.NODE_ENV === 'production' ? 'error' : 'warn'
+    'no-unused-vars': import.meta.env.MODE === 'production' ? 'error' : 'warn'
   }
 }
