@@ -25,6 +25,8 @@ import { useRoute, useRouter } from 'vue-router'
 import TheNavbar from '@/components/TheNavbar.vue'
 import { useUsersStore } from '@/stores/UsersStore'
 import NProgress from 'nprogress'
+import 'nprogress/nprogress.css'
+import { useHead } from '@unhead/vue'
 
 const usersStore = useUsersStore()
 usersStore.fetchAuthUser()
@@ -41,11 +43,26 @@ useRouter().beforeEach(() => {
 function pageReady () {
   NProgress.done()
 }
+
+useHead({
+  meta: [
+    { name: 'description', content: 'A Vue3 Forum SPA written with the composition API and Firebase 🔥' },
+
+    {property: "og:title", content: "Vue3 Masterclass Forum"},
+    {property: "og:description", content: "A Vue3 Forum SPA written with the composition API and Firebase 🔥"},
+    {property: "og:image", content: "https://vueschool.io/media/f007f6057444d9a7f567163391d2b366/vuejs-3-master-class-not-transparent.jpg"},
+
+    {name: "twitter:title", content: "Vue3 Masterclass Forum"},
+    {name: "twitter:description", content: "A Vue3 Forum SPA written with the composition API and Firebase 🔥"},
+    {name: "twitter:image", content: "https://vueschool.io/media/f007f6057444d9a7f567163391d2b366/vuejs-3-master-class-not-transparent.jpg"},
+    {name: "twitter:card", content: "summary_large_image"},
+  ]
+})
 </script>
 
 <style>
 @import "assets/style.css";
-@import "~nprogress/nprogress.css";
+@import "nprogress/nprogress.css";
 #nprogress .bar{
   background: #57AD8D !important
 }
