@@ -11,10 +11,11 @@
 
 <script setup>
 
-import { defineProps, computed } from 'vue'
+import { defineProps, computed, ref } from 'vue'
 import CategoryListItem from '@/components/CategoryListItem.vue'
 import { useCategoriesStore } from '@/stores/CategoriesStore'
 import { useForumsStore } from '../stores/ForumsStore'
+import { useHead } from '@unhead/vue'
 
 const categoriesStore = useCategoriesStore()
 
@@ -36,6 +37,14 @@ const category = computed(() => {
   return categoriesStore.categoryById(props.id)
 })
 
+const title = ref(`${category.value?.name} | Vue3 Masterclass Forum`)
+useHead({
+  title: () => title.value,
+  meta: [
+  {property: "og:title", content: () => title.value},
+  {name: "twitter:title", content: () => title.value}
+   ]
+})
 </script>
 
 <style scoped>
