@@ -8,25 +8,25 @@ const routes = [
   {
     path: '/',
     name: 'Home',
-    component: () => import(/* webpackChunkName: "Home" */'@/pages/HomeIndex')
+    component: () => import('@/pages/HomeIndex')
   },
   {
     path: '/me',
     name: 'Profile',
-    component: () => import(/* webpackChunkName: "Profile" */'@/pages/ProfileShow.vue'),
+    component: () => import('@/pages/ProfileShow.vue'),
     meta: { toTop: true, smoothScroll: true, requiresAuth: true }
   },
   {
     path: '/me/edit',
     name: 'ProfileEdit',
-    component: () => import(/* webpackChunkName: "Profile Edit" */'@/pages/ProfileShow.vue'),
+    component: () => import('@/pages/ProfileShow.vue'),
     props: { edit: true },
     meta: { requiresAuth: true }
   },
   {
     path: '/category/:id',
     name: 'CategoryShow',
-    component: () => import(/* webpackChunkName: "Category" */'@/pages/CategoryShow'),
+    component: () => import('@/pages/CategoryShow'),
     props: true,
     async beforeEnter (to, from, next) {
       const redirect = await checkResourceExistsGuard(useCategoriesStore().categories, 'categories', '🏷', to)
@@ -40,7 +40,7 @@ const routes = [
   {
     path: '/forum/:id',
     name: 'ForumShow',
-    component: () => import(/* webpackChunkName: "Forum" */'@/pages/ForumShow'),
+    component: () => import('@/pages/ForumShow'),
     props: true,
     async beforeEnter (to, from, next) {
       const redirect = await checkResourceExistsGuard(useForumsStore().forums, 'forums', '🗣', to)
@@ -54,14 +54,14 @@ const routes = [
   {
     path: '/forum/:forumId/thread/create',
     name: 'ThreadCreate',
-    component: () => import(/* webpackChunkName: "ThreadCreate" */'@/pages/ThreadCreate'),
+    component: () => import('@/pages/ThreadCreate'),
     props: true,
     meta: { requiresAuth: true }
   },
   {
     path: '/thread/:id',
     name: 'ThreadShow',
-    component: () => import(/* webpackChunkName: "Thread" */'@/pages/ThreadShow'),
+    component: () => import('@/pages/ThreadShow'),
     props: true,
     async beforeEnter (to, from, next) {
       const redirect = await checkResourceExistsGuard(useThreadsStore().threads, 'threads', '📄', to)
@@ -75,7 +75,7 @@ const routes = [
   {
     path: '/thread/:id/edit',
     name: 'ThreadEdit',
-    component: () => import(/* webpackChunkName: "ThreadEdit" */'@/pages/ThreadEdit'),
+    component: () => import('@/pages/ThreadEdit'),
     props: true,
     // meta: { requiresAuth: true },
     async beforeEnter (to, from, next) {
@@ -99,13 +99,13 @@ const routes = [
   {
     path: '/register',
     name: 'Register',
-    component: () => import(/* webpackChunkName: "Register" */'@/pages/RegisterShow'),
+    component: () => import('@/pages/RegisterShow'),
     meta: { requiresGuest: true }
   },
   {
     path: '/signin',
     name: 'SignIn',
-    component: () => import(/* webpackChunkName: "SignIn" */'@/pages/SignInShow'),
+    component: () => import('@/pages/SignInShow'),
     meta: { requiresGuest: true }
   },
   {
@@ -119,12 +119,12 @@ const routes = [
   {
     path: '/signedout',
     name: 'SignedOut',
-    component: () => import(/* webpackChunkName: "SignedOut" */'@/pages/SignedOutShow')
+    component: () => import('@/pages/SignedOutShow')
   },
   {
     path: '/:pathMatch(.*)*',
     name: '404',
-    component: () => import(/* webpackChunkName: "404" */'@/pages/404NotFound')
+    component: () => import('@/pages/404NotFound')
   }
 ]
 
